Copying
//UE4/UT-Staging/...

to //depot/UE4-UT/...

[CL 3104597 by Peter Knepley in UT-Clean branch]

+++ Engine/Source/Runtime/UnrealAudio/Private/UnrealAudioDeviceFormat.cpp
@@ -99,7 +99,7 @@ namespace UAudio
 			{
 				for (uint32 ChannelIndex = 0; ChannelIndex < ConvertInfo.NumChannels; ++ChannelIndex)
 				{
-					int32 Value = FromBufferPtr[ChannelIndex];
+					int64 Value = FromBufferPtr[ChannelIndex];
 					ToBufferPtr[ChannelIndex] = (ToType)(Value >> ShiftAmount);
 				}
 				FromBufferPtr += ConvertInfo.FromChannels;
@@ -112,7 +112,7 @@ namespace UAudio
 			{
 				for (uint32 ChannelIndex = 0; ChannelIndex < ConvertInfo.NumChannels; ++ChannelIndex)
 				{
-					int32 Value = FromBufferPtr[ChannelIndex];
+					int64 Value = FromBufferPtr[ChannelIndex];
 					ToBufferPtr[ChannelIndex] = (ToType)(Value << ShiftAmount);
 				}
 				FromBufferPtr += ConvertInfo.FromChannels;
